+++ kruskals_algorithm/Cargo.toml
@@ -1,18 +1,7 @@
-[package]
-name = "kruskals_algorithm"
-version = "0.1.0"
-authors = ["gcfbn <bmekarski@interia.pl>"]
-edition = "2018"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0"
-thiserror = "1.0.25"
-parse-display = "0.5.0"
-
-[dev-dependencies]
-test-case = "1.1.0"
-
-[dependencies.clap]
-version = "~3.0.0-beta.2"
+members = [
+    "graph",
+    "kruskal",
+    "runner"
+]
