+++ kruskals_algorithm/graph/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "graph"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+thiserror = "1.0.25"
+parse-display = "0.5.0"
+
+[dev-dependencies]
+test-case = "1.1.0"
