+++ kruskals_algorithm/graph/src/lib.rs
@@ -1,0 +1,19 @@
+// extern these crates only when running tests
+#[cfg(test)]
+extern crate test_case;
+
+mod dfs;
+mod errors;
+mod reader;
+mod structures;
+
+pub use crate::errors::{AddingEdgeError, BuildGraphError, GraphParametersParsingError, ParsingEdgeError, Result};
+pub use crate::reader::build_graph_from_file;
+pub use crate::structures::{Edge, EdgeDescription, Graph};
+
+// pub fn build_graph<P>(filename: P) -> Result<Graph>
+// where
+//     P: AsRef<Path>,
+// {
+//     build_graph_from_file(filename)
+// }
