+++ kruskals_algorithm/graph/tests/data/edge_errors/error_adding_edge_wrong_from_index.txt
