+++ kruskals_algorithm/graph/tests/data/edge_errors/error_parsing_edge_non_integer_weight.txt
