+++ kruskals_algorithm/graph/tests/data/graph_building_errors/error_graph_not_connected.txt
