+++ kruskals_algorithm/kruskal/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "kruskal"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+graph = { path = "../graph" }
+
+[dev-dependencies]
+test-case = "1.1.0"
