+++ kruskals_algorithm/kruskal/src/lib.rs
@@ -1,0 +1,7 @@
+// extern these crates only when running tests
+#[cfg(test)]
+extern crate test_case;
+
+mod algorithm;
+
+pub use algorithm::calculate_min_total_weight;
