+++ kruskals_algorithm/kruskal/tests/data/passing1.txt
