+++ kruskals_algorithm/kruskal/tests/data/passing2.txt
