+++ kruskals_algorithm/kruskal/tests/data/passing3.txt
