+++ kruskals_algorithm/kruskal/tests/data/passing6.txt
