+++ kruskals_algorithm/kruskal/tests/data/passing7.txt
