+++ kruskals_algorithm/runner/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "runner"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+thiserror = "1.0.25"
+kruskal = { path = "../kruskal" }
+graph = { path = "../graph" }
+
+[dependencies.clap]
+version = "~3.0.0-beta.2"
