--- kruskals_algorithm/src/data/mod.rs
@@ -1,6 +1,0 @@
-mod dfs;
-mod reader;
-mod structures;
-
-pub use reader::build_graph_from_file;
-pub use structures::{Edge, EdgeDescription, Graph};
